requirements.txt
python-telegram-bot
graia-application-mirai
aiofiles
pydantic==1.7.1
coloredlogs
pysocks
python-magic
pillow